Final merge out-of-band development
This is the final commit that brings all the development done in a seperate git
repository into the git repository of github.
All previous commits in this branch since the 'add readme.md' commit were done
in a repo that had nothing in common with the github repo.
So a lot of cherry-picking got most patches imported but some code slipped
trough the mazes, this commit fixes those last differences.